Updated "updateItem" function

+++ controller/ItemController.py
@@ -33,6 +33,13 @@ class ItemController:
         else:
             return jsonify('ID Not Found'), 405
 
+    def getDictByID(self, id):
+        daoRes = self.dao.getByID(id)
+        if daoRes:
+            return self.dictionary(daoRes[0])
+        else:
+            return {}
+
     def addNewItem(self, json):
         i_name = json['i_name']
         i_category = json['i_category']
@@ -44,25 +51,29 @@ class ItemController:
         return jsonify(json), 201
 
     def deleteItem(self, id):
-        self.dao.deleteItemByID(id)
-        return jsonify("Item #%s deleted" % id)
+        daoRes = self.dao.getByID(id)
+        if daoRes:
+            self.dao.deleteItemByID(id)
+            return jsonify(self.dictionary(daoRes[0]))
+        else:
+            return jsonify('ID Not Found'), 405
 
     def updateItem(self, id, reqjson):
 
-        # print(self.getByID(id).json()['i_name'])
-        # i_name = oldjson['i_name']
-        # i_category = oldjson['i_category']
-        # i_stock = oldjson['i_stock']
-        # i_price = oldjson['i_price']
+        oldjson = self.getDictByID(id)
+        i_name = oldjson['i_name']
+        i_category = oldjson['i_category']
+        i_stock = oldjson['i_stock']
+        i_price = oldjson['i_price']
 
-    # if json['i_name']:
-        i_name = reqjson['i_name']
-    # if json['i_category']:
-        i_category = reqjson['i_category']
-    # if json['i_stock']:
-        i_stock = reqjson['i_stock']
-    # if json['i_price']:
-        i_price = reqjson['i_price']
+        if reqjson['i_name'] != '':
+            i_name = reqjson['i_name']
+        if reqjson['i_category'] != '':
+            i_category = reqjson['i_category']
+        if reqjson['i_stock'] != '':
+            i_stock = reqjson['i_stock']
+        if reqjson['i_price'] != '':
+            i_price = reqjson['i_price']
 
         daoRes = self.dao.updateItemByID(id, i_name, i_category, i_stock, i_price)
         if daoRes:
